backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
openai==1.30.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
huggingface_hub>=0.23.0
Pillow>=10.0.0
